requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==46.0.2
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
playwright==1.48.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
pymongo==4.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.2
requests==2.32.3
rsa==4.9.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0